chore: rename spline file to be more clear
